Added "Code Golfing Considered Harmful".
"Code Golfing Considered Harmful" is an article I wrote when I first joined the Shadertoy community. It is a stern warning not to get lured into thinking that code golfing is a thing outside of the community of code golfing. In all real world scenarios it's directly harmful to your code and to you as a programmer.